Use index from u2fdevice struct
`index` was uninitialized and happened to take on the value
0 on my machine in this case. This meant that `u2fh_devs_discover`
did the right thing if the device was never removed and plugged
back in. If it did get removed and reinserted, each call to
`u2fh_devs_discover` would increment the max_index because
`ping_device` for `index=0` would always fail.